+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.2.2 - 2025-05-16
+- Limit search output to 500 lines to prevent overly large responses.
+
 ## 0.2.1 - 2025-05-16
 - Limit search to 10 items by default.